feat: generate new point template
