+++ src/presenter/trips-presenter.js
@@ -1,10 +1,10 @@
 import EditPointView from '../view/edit-point-view.js';
-//import NewPointView from '../view/add-new-point-view.js';
+import NewPointView from '../view/add-new-point-view.js';
 import PointView from '../view/point-view.js';
 import PointListView from '../view/point-list-view.js';
 import ListEmptyView from '../view/list-empty-view.js';
 import { isEscapeKey } from '../util.js';
-import { render, replace} from '../framework/render.js';
+import { render, replace, RenderPosition} from '../framework/render.js';
 
 
 export default class TripPresenter {
@@ -33,7 +33,7 @@ export default class TripPresenter {
 
 
     render(this.#pointListComponent, this.#pointsContainer);
-    //render(new NewPointView(), this.#pointListComponent.element, RenderPosition.AFTERBEGIN);
+    render(new NewPointView(), this.#pointListComponent.element, RenderPosition.AFTERBEGIN);
 
     this.#listPoints.forEach((point) => this.#renderPoint(point));
   }
