+++ src/view/add-new-point-view.js
@@ -2,7 +2,7 @@ import AbstractView from '../framework/view/abstract-view.js';
 import { destinations, offersByType } from '../mock/point.js';
 import dayjs from 'dayjs';
 import { DATE_FORMAT } from '../const.js';
-//import { DefaultNewPoint } from '../mock/new-point.js';
+import { DefaultNewPoint } from '../mock/const.js';
 
 
 const createNewPointTemplate = (point) => {
@@ -55,87 +55,94 @@ const createNewPointTemplate = (point) => {
     }
     return template;
   };
-  //console.log(pointDestination);
+
+
+  const picturesTemplate = () => {
+    let template = '';
+    if (pointDestination) {
+      template = pointDestination.pictures
+        .map((picture) => `<img class="event__photo" src="${picture.src}" alt="${picture.description}">`)
+        .join('');
+    }
+    return template;
+  };
   return (
     `<li class="trip-events__item">
-    <form class="event event--edit" action="#" method="post">
-      <header class="event__header">
-        <div class="event__type-wrapper">
-          <label class="event__type  event__type-btn" for="event-type-toggle-${id}">
-            <span class="visually-hidden">Choose event type</span>
-            <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
-          </label>
-          <input class="event__type-toggle  visually-hidden" id="event-type-toggle-${id}" type="checkbox">
+      <form class="event event--edit" action="#" method="post">
+        <header class="event__header">
+          <div class="event__type-wrapper">
+            <label class="event__type  event__type-btn" for="event-type-toggle-${id}">
+              <span class="visually-hidden">Choose event type</span>
+              <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
+            </label>
+            <input class="event__type-toggle  visually-hidden" id="event-type-toggle-${id}" type="checkbox">
 
-          <div class="event__type-list">
-            <fieldset class="event__type-group">
-              <legend class="visually-hidden">Event type</legend>
-                ${tripOptionsList}
-            </fieldset>
-          </div>
-        </div>
-
-        <div class="event__field-group  event__field-group--destination">
-          <label class="event__label  event__type-output" for="event-destination-${id}">
-            ${type}
-          </label>
-          <input class="event__input  event__input--destination" id="event-destination-${id}" type="text" name="event-destination" value="${pointDestination.name}" list="destination-list-${id}">
-          <datalist id="destination-list-${id}">
-            ${destinationToChoose}
-          </datalist>
-        </div>
-
-        <div class="event__field-group  event__field-group--time">
-          <label class="visually-hidden" for="event-start-time-${id}">From</label>
-          <input class="event__input  event__input--time" id="event-start-time-${id}" type="text" name="event-start-time" value="${parseDateStart.format(DATE_FORMAT)}">
-          &mdash;
-          <label class="visually-hidden" for="event-end-time-1">To</label>
-          <input class="event__input  event__input--time" id="event-end-time-${id}" type="text" name="event-end-time" value="${parseDateEnd.format(DATE_FORMAT)}">
-        </div>
-
-        <div class="event__field-group  event__field-group--price">
-          <label class="event__label" for="event-price-${id}">
-            <span class="visually-hidden">Price</span>
-            &euro;
-          </label>
-          <input class="event__input  event__input--price" id="event-price-${id}" type="text" name="event-price" value="${basePrice}">
-        </div>
-
-        <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-        <button class="event__reset-btn" type="reset">Cancel</button>
-      </header>
-      <section class="event__details">
-        ${offersSectionTemplate()}
-
-        <section class="event__section  event__section--destination">
-          <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-          <p class="event__destination-description">${pointDestination.description}</p>
-
-          <div class="event__photos-container">
-            <div class="event__photos-tape">
-              <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
-              <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
-              <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
-              <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
-              <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
+            <div class="event__type-list">
+              <fieldset class="event__type-group">
+                <legend class="visually-hidden">Event type</legend>
+                  ${tripOptionsList}
+              </fieldset>
             </div>
           </div>
+
+          <div class="event__field-group  event__field-group--destination">
+            <label class="event__label  event__type-output" for="event-destination-${id}">
+              ${type}
+            </label>
+            <input class="event__input  event__input--destination" id="event-destination-${id}" type="text" name="event-destination" value="${pointDestination ? pointDestination.name : ''}" list="destination-list-${id}">
+            <datalist id="destination-list-${id}">
+              ${destinationToChoose}
+            </datalist>
+          </div>
+
+          <div class="event__field-group  event__field-group--time">
+            <label class="visually-hidden" for="event-start-time-${id}">From</label>
+            <input class="event__input  event__input--time" id="event-start-time-${id}" type="text" name="event-start-time" value="${parseDateStart.format(DATE_FORMAT)}">
+            &mdash;
+            <label class="visually-hidden" for="event-end-time-1">To</label>
+            <input class="event__input  event__input--time" id="event-end-time-${id}" type="text" name="event-end-time" value="${parseDateEnd.format(DATE_FORMAT)}">
+          </div>
+
+          <div class="event__field-group  event__field-group--price">
+            <label class="event__label" for="event-price-${id}">
+              <span class="visually-hidden">Price</span>
+              &euro;
+            </label>
+            <input class="event__input  event__input--price" id="event-price-${id}" type="text" name="event-price" value="${basePrice}">
+          </div>
+
+          <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
+          <button class="event__reset-btn" type="reset">Cancel</button>
+        </header>
+        <section class="event__details">
+          ${offersSectionTemplate()}
+
+          <section class="event__section  event__section--destination">
+            <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+            <p class="event__destination-description">${pointDestination ? pointDestination.description : ''}</p>
+
+            <div class="event__photos-container">
+              <div class="event__photos-tape">
+                ${picturesTemplate()}
+              </div>
+            </div>
+          </section>
         </section>
-      </section>
-    </form>
-  </li>`);
+      </form>
+    </li>`
+  );
 };
 
 export default class NewPointView extends AbstractView {
-  // #point = null;
+  #point = null;
 
-  // constructor({point = DefaultNewPoint}) {
-  //   super();
-  //   this.#point = point;
-  // }
+  constructor(point = DefaultNewPoint) {
+    super();
+    this.#point = point;
+  }
 
   get template() {
-    return createNewPointTemplate();
+    return createNewPointTemplate(this.#point);
   }
 
 }
